ウィジェット編集のジェスチャー不具合を修正し、配置グリッドを3倍細かく変更
- APP LISTのS/M/Lサイズ変更時、Sサイズはアイコンのみ表示（アプリ名を省略）
- ウィジェット編集モードを四隅の太枠ブラケットでどこからでもリサイズできるようにし、
  移動はヘッダーなどの長押し→ドラッグに変更
- タップ（短押しで編集モード終了）と長押し→ドラッグ（移動）を1つのジェスチャー
  ループに統合し、以下の不具合を解消:
  - 動きのない長押しが常にタップ扱いになり編集モードに入れない不具合
    （長押し成立の判定より先に「指が離れた」判定をしていた順序ミス）
  - イベント未消費により、編集モードに入った直後に祖先の背景タップ検出が
    同じジェスチャーを二重処理し、即座に編集モードが解除される不具合
- ウィジェットキャンバスの全体グリッドを3倍細かくし（例: 横画面 4×4→12×12）、
  APP LIST以外のウィジェットも含めて移動・リサイズの自由度を大幅に向上
  （保存レイアウトの座標の意味が変わるため、保存キーをv2に分離し、
  既存ユーザーは一度だけ新グリッドでのデフォルト配置に戻る）

+++ app/src/main/java/com/example/girdlauncher/model/PlacedWidget.kt
@@ -4,16 +4,17 @@ package com.example.girdlauncher.model
  * ホーム画面のウィジェットキャンバス上に配置された1つのウィジェット。
  *
  * @property type ウィジェットの種類。
- * @property col 配置されている列（0始まり）。
- * @property row 配置されている行（0始まり）。
+ * @property col 配置されている列（0始まり）。左端・上端のコーナーハンドルでリサイズすると
+ *   小数値を取り得る。
+ * @property row 配置されている行（0始まり）。[col]と同様に小数値を取り得る。
  * @property colSpan 横方向に占めるセル数。ウィジェットの種類によっては、アイコンサイズ単位での
  *   細かいリサイズに対応するため小数値を取り得る。
  * @property rowSpan 縦方向に占めるセル数。[colSpan]と同様に小数値を取り得る。
  */
 data class PlacedWidget(
     val type: WidgetPanel,
-    val col: Int,
-    val row: Int,
+    val col: Float,
+    val row: Float,
     val colSpan: Float,
     val rowSpan: Float
 ) {
